Post-verification fixes, merge with colins constants file
