+++ src/app/calculator/compressed-air/compressed-air-constants.ts
@@ -1,82 +1,173 @@
-export const pipeSizeConstants: Array<{ value: number, display: string }> = [
+export const standardSizesConstant: Array<{ size: string, display: string }> = [
     {
-        value: 0,
+        size: 'CUSTOM',
+        display: 'Use Custom Size'
+    },
+    {
+        size: 'oneHalf',
         display: '1/2'
     },
     {
-        value: 1,
+        size: 'threeFourths',
         display: '3/4'
     },
     {
-        value: 2,
+        size: 'one',
         display: '1'
     },
     {
-        value: 3,
+        size: 'oneAndOneFourth',
         display: '1 1/4'
     },
     {
-        value: 4,
+        size: 'oneAndOneHalf',
         display: '1 1/2'
     },
     {
-        value: 5,
+        size: 'two',
         display: '2'
     },
     {
-        value: 6,
+        size: 'twoAndOneHalf',
         display: '2 1/2'
     },
     {
-        value: 7,
+        size: 'three',
         display: '3'
     },
     {
-        value: 8,
+        size: 'threeAndOneHalf',
         display: '3 1/2'
     },
     {
-        value: 9,
+        size: 'four',
         display: '4'
     },
     {
-        value: 10,
+        size: 'five',
         display: '5'
     },
     {
-        value: 11,
+        size: 'six',
         display: '6'
     },
     {
-        value: 12,
+        size: 'eight',
         display: '8'
     },
     {
-        value: 13,
+        size: 'ten',
         display: '10'
     },
     {
-        value: 14,
+        size: 'twelve',
         display: '12'
     },
     {
-        value: 15,
+        size: 'fourteen',
         display: '14'
     },
     {
-        value: 16,
+        size: 'sixteen',
         display: '16'
     },
     {
-        value: 17,
+        size: 'eighteen',
         display: '18'
     },
     {
-        value: 18,
+        size: 'twenty',
         display: '20'
     },
     {
-        value: 19,
+        size: 'twentyFour',
         display: '24'
-    }
-]
+    },
+];
+
+export const metricSizesConstant: Array<{ size: string, display: string }> = [
+    {
+        size: 'CUSTOM',
+        display: 'Use Custom Size'
+    },
+    {
+        size: 'oneHalf',
+        display: '15'
+    },
+    {
+        size: 'threeFourths',
+        display: '20'
+    },
+    {
+        size: 'one',
+        display: '25'
+    },
+    {
+        size: 'oneAndOneFourth',
+        display: '32'
+    },
+    {
+        size: 'oneAndOneHalf',
+        display: '40'
+    },
+    {
+        size: 'two',
+        display: '50'
+    },
+    {
+        size: 'twoAndOneHalf',
+        display: '65'
+    },
+    {
+        size: 'three',
+        display: '80'
+    },
+    {
+        size: 'threeAndOneHalf',
+        display: '90'
+    },
+    {
+        size: 'four',
+        display: '100'
+    },
+    {
+        size: 'five',
+        display: '125'
+    },
+    {
+        size: 'six',
+        display: '150'
+    },
+    {
+        size: 'eight',
+        display: '200'
+    },
+    {
+        size: 'ten',
+        display: '250'
+    },
+    {
+        size: 'twelve',
+        display: '300'
+    },
+    {
+        size: 'fourteen',
+        display: '350'
+    },
+    {
+        size: 'sixteen',
+        display: '400'
+    },
+    {
+        size: 'eighteen',
+        display: '450'
+    },
+    {
+        size: 'twenty',
+        display: '500'
+    },
+    {
+        size: 'twentyFour',
+        display: '600'
+    },
+];
