--- src/app/calculator/compressed-air/compressedAirConstants.ts
@@ -1,173 +1,0 @@
-export const standardSizesConstant: Array<{ size: string, display: string }> = [
-    {
-        size: 'CUSTOM',
-        display: 'Use Custom Size'
-    },
-    {
-        size: 'oneHalf',
-        display: '1/2'
-    },
-    {
-        size: 'threeFourths',
-        display: '3/4'
-    },
-    {
-        size: 'one',
-        display: '1'
-    },
-    {
-        size: 'oneAndOneFourth',
-        display: '1 1/4'
-    },
-    {
-        size: 'oneAndOneHalf',
-        display: '1 1/2'
-    },
-    {
-        size: 'two',
-        display: '2'
-    },
-    {
-        size: 'twoAndOneHalf',
-        display: '2 1/2'
-    },
-    {
-        size: 'three',
-        display: '3'
-    },
-    {
-        size: 'threeAndOneHalf',
-        display: '3 1/2'
-    },
-    {
-        size: 'four',
-        display: '4'
-    },
-    {
-        size: 'five',
-        display: '5'
-    },
-    {
-        size: 'six',
-        display: '6'
-    },
-    {
-        size: 'eight',
-        display: '8'
-    },
-    {
-        size: 'ten',
-        display: '10'
-    },
-    {
-        size: 'twelve',
-        display: '12'
-    },
-    {
-        size: 'fourteen',
-        display: '14'
-    },
-    {
-        size: 'sixteen',
-        display: '16'
-    },
-    {
-        size: 'eighteen',
-        display: '18'
-    },
-    {
-        size: 'twenty',
-        display: '20'
-    },
-    {
-        size: 'twentyFour',
-        display: '24'
-    },
-];
-
-export const metricSizesConstant: Array<{ size: string, display: string }> = [
-    {
-        size: 'CUSTOM',
-        display: 'Use Custom Size'
-    },
-    {
-        size: 'oneHalf',
-        display: '15'
-    },
-    {
-        size: 'threeFourths',
-        display: '20'
-    },
-    {
-        size: 'one',
-        display: '25'
-    },
-    {
-        size: 'oneAndOneFourth',
-        display: '32'
-    },
-    {
-        size: 'oneAndOneHalf',
-        display: '40'
-    },
-    {
-        size: 'two',
-        display: '50'
-    },
-    {
-        size: 'twoAndOneHalf',
-        display: '65'
-    },
-    {
-        size: 'three',
-        display: '80'
-    },
-    {
-        size: 'threeAndOneHalf',
-        display: '90'
-    },
-    {
-        size: 'four',
-        display: '100'
-    },
-    {
-        size: 'five',
-        display: '125'
-    },
-    {
-        size: 'six',
-        display: '150'
-    },
-    {
-        size: 'eight',
-        display: '200'
-    },
-    {
-        size: 'ten',
-        display: '250'
-    },
-    {
-        size: 'twelve',
-        display: '300'
-    },
-    {
-        size: 'fourteen',
-        display: '350'
-    },
-    {
-        size: 'sixteen',
-        display: '400'
-    },
-    {
-        size: 'eighteen',
-        display: '450'
-    },
-    {
-        size: 'twenty',
-        display: '500'
-    },
-    {
-        size: 'twentyFour',
-        display: '600'
-    },
-];
