+++ src/app/calculator/compressed-air/system-capacity/system-capacity.component.html
@@ -66,7 +66,7 @@
               Select a pipe size from the list of nominal schedule 40 pipe inner diameters or enter a custom pipe size (inner diameter). For a quicker estimation,
               ignore all pipe sizes 
               <span *ngIf="settings.unitsOfMeasure != 'Metric'">2"</span>
-              <span *ngIf="settings.unitsOfMeasure == 'Metric'">50cm</span>  
+              <span *ngIf="settings.unitsOfMeasure == 'Metric'">50mm</span>  
               or less (they do not add significant system capacity).
               <hr>
             </small>
